services/openai/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

fastapi==0.118.0
openai==2.1.0
pydantic==2.11.10
python-dotenv==1.1.1
uvicorn==0.37.0
